+++ src/simulation/elements/FRAY.cpp
@@ -69,8 +69,8 @@ int Element_FRAY::update(UPDATE_FUNC_ARGS)
 						if (!r)
 							r = sim->photons[y+nyi+nyy][x+nxi+nxx];
 						if (r && !(sim->elements[TYP(r)].Properties & TYPE_SOLID)){
-							parts[ID(r)].vx += nxi*((parts[i].temp-273.15)/10.0f);
-							parts[ID(r)].vy += nyi*((parts[i].temp-273.15)/10.0f);
+							parts[ID(r)].vx += nxi*(((float)parts[i].temp-273.15)/10.0f);
+							parts[ID(r)].vy += nyi*(((float)parts[i].temp-273.15)/10.0f);
 						}
 					}
 				}
