+++ src/simulation/elements/PRTO.cpp
@@ -58,7 +58,7 @@ int Element_PRTO::update(UPDATE_FUNC_ARGS)
 {
 	int r, nnx, rx, ry, np, fe = 0;
 	int count = 0;
-	parts[i].tmp = (int)((parts[i].temp-73.15f)/100+1);
+	parts[i].tmp = (int)(((float)parts[i].temp-73.15f)/100+1);
 	if (parts[i].tmp>=CHANNELS) parts[i].tmp = CHANNELS-1;
 	else if (parts[i].tmp<0) parts[i].tmp = 0;
 	for (count=0; count<8; count++)
